[refactor] ndpp::Tensor Optimization.
- Data reference: In TensorDevice::DeviceRefer(), it use reference on _shape & _strides, and adding other judgement conditions of data type whether data types are correct.
- Memory deallocation: In TensorDevice::DeviceDeAlloc(), Adjusting the condition whether to do memory destroy.
- Memory allocation: In TensorDevice::DeviceAlloc(), Adding judgement whether _data is nullptr after memory allocation.
- Migraing the function getInfoTyByDataTy() in namespace ndpp::ndpp_tensor_base.

[refactor] Adding ndpp::Integer.

- Defined in core/include/base/BasicType.hpp.
- Supporting int64_t or int32_t, which based on environment.
- Using ndpp::IntArray instead of ndpp::SizeTArray, and replaceing all functions which are using ndpp::SizeTArray.

+++ core/include/tensor/GpuTensorArithmetic.cuh
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <include/base/BasicType.hpp>
 #include <include/base/ScalarType.hpp>
 #include <include/tensor/TensorArithmetic.hpp>
 #include <include/tensor/TensorArithmeticProc.hpp>
@@ -12,8 +13,8 @@ namespace ndpp_arithmetic
 {
 
 void gpuArithForwardKernel(const void *a, const void *b, void *c, const ndpp_memory::ScalarType a_stype, const ndpp_memory::ScalarType b_stype,
-                           const size_t *a_allocated_shape, const size_t *a_acutal_shape, const size_t *a_strides, int64_t a_shape_dim, int64_t a_total,
-                           const size_t *b_allocated_shape, const size_t *b_acutal_shape, const size_t *b_strides, int64_t b_shape_dim, int64_t b_total,
+                           const Integer *a_allocated_shape, const Integer *a_acutal_shape, const Integer *a_strides, int64_t a_shape_dim, int64_t a_total,
+                           const Integer *b_allocated_shape, const Integer *b_acutal_shape, const Integer *b_strides, int64_t b_shape_dim, int64_t b_total,
                            int64_t loop_total, Arithmetic arith_type);
 
 }; // namespace ndpp_arithmetic
